requirements.txt
openai>=1.30.0
python-dotenv>=0.19.0
pandas>=1.5.0,<2.0.0
numpy>=1.21.0,<2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.60.0
colorama>=0.4.4 